requirements.txt
SQLAlchemy[asyncio]>=1.4.49
langchain-core
langchain_community 
langgraph 
langchain_chroma
chromadb
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
beautifulsoup4
python-dotenv
colorama
langserve
sse_starlette
uvicorn
gunicorn
fastapi
openai
langchain-openai
streamlit
